+++ src/lib/simulations/contextConditionals.ts
@@ -23,7 +23,7 @@ export function initializeContextConditionals(context: OptimizerContext) {
     if (context.teammate2Metadata?.characterId) calculateTeammateDynamicConditionals(action, context.teammate2Metadata, 2)
 
     // Reconstruct arrays after transfer
-    action.precomputedStats.a = new Float32Array(Object.values(action.precomputedStats.a))
+    action.precomputedStats.a = new Float64Array(Object.values(action.precomputedStats.a))
 
     // Rebuild entityRegistry from entitiesArray after serialization
     if (action.config) {
