+++ src/lib/simulations/statSimulationTypes.ts
@@ -89,7 +89,7 @@ export type SimulateBuildResult = {
 
 export type RunStatSimulationsResult = {
   x: ComputedStatsContainer,
-  xa: Float32Array,
+  xa: Float64Array,
   ca: Float32Array,
   simScore: number,
   key?: string,
